requirements.txt
Django==1.4.2
South==0.7.6
django-cms
django-sekizai
django-mptt
django-classy-tags
django-autoslug
django-debug-toolbar
django-bootstrap-toolkit
django-registration
django-celery
django-celery-with-mongodb
pytils
PIL
pymongo


#celery
#celery-with-mongodb
